+++ NEWS.md
@@ -10,6 +10,13 @@
 * `update`s agora retornam pelo construtor interno, conferindo mais robustez (#10)
 * Objetos `modprev` agora podem carregar atributos genéricos pertinentes ao modelo estimado (#11)
 
+# janelamovel
+
+## New features
+
+* `janelamovel` agora tem argumento `full.output`, um booleano indicando o nivel de complexidade da
+  saida. Se true, retorna alem da previsao o modelo e variaveis explicativas utilizados
+
 # modprev 1.4.5
 
 ## New features